social_media_app/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
motor==3.3.2
pymongo>=4.0
dnspython==2.6.1
email-validator==2.1.1
google-generativeai==0.3.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
apscheduler==3.10.4
python-dotenv==1.0.1
requests==2.31.0
